Refactor to move listener related code to package listener
