+++ core/src/test/java/com/datastrato/gravitino/listener/api/event/TestCatalogEvent.java
@@ -14,8 +14,8 @@ import com.datastrato.gravitino.CatalogChange;
 import com.datastrato.gravitino.NameIdentifier;
 import com.datastrato.gravitino.Namespace;
 import com.datastrato.gravitino.catalog.CatalogDispatcher;
-import com.datastrato.gravitino.catalog.CatalogEventDispatcher;
 import com.datastrato.gravitino.exceptions.GravitinoRuntimeException;
+import com.datastrato.gravitino.listener.CatalogEventDispatcher;
 import com.datastrato.gravitino.listener.DummyEventListener;
 import com.datastrato.gravitino.listener.EventBus;
 import com.datastrato.gravitino.listener.api.info.CatalogInfo;
