+++ core/src/test/java/com/datastrato/gravitino/listener/api/event/TestFilesetEvent.java
@@ -12,12 +12,12 @@ import static org.mockito.Mockito.when;
 import com.datastrato.gravitino.NameIdentifier;
 import com.datastrato.gravitino.Namespace;
 import com.datastrato.gravitino.catalog.FilesetDispatcher;
-import com.datastrato.gravitino.catalog.FilesetEventDispatcher;
 import com.datastrato.gravitino.exceptions.GravitinoRuntimeException;
 import com.datastrato.gravitino.file.Fileset;
 import com.datastrato.gravitino.file.FilesetChange;
 import com.datastrato.gravitino.listener.DummyEventListener;
 import com.datastrato.gravitino.listener.EventBus;
+import com.datastrato.gravitino.listener.FilesetEventDispatcher;
 import com.datastrato.gravitino.listener.api.info.FilesetInfo;
 import com.google.common.collect.ImmutableMap;
 import java.util.Arrays;
