+++ core/src/test/java/com/datastrato/gravitino/listener/api/event/TestMetalakeEvent.java
@@ -15,9 +15,9 @@ import com.datastrato.gravitino.NameIdentifier;
 import com.datastrato.gravitino.exceptions.GravitinoRuntimeException;
 import com.datastrato.gravitino.listener.DummyEventListener;
 import com.datastrato.gravitino.listener.EventBus;
+import com.datastrato.gravitino.listener.MetalakeEventDispatcher;
 import com.datastrato.gravitino.listener.api.info.MetalakeInfo;
 import com.datastrato.gravitino.metalake.MetalakeDispatcher;
-import com.datastrato.gravitino.metalake.MetalakeEventDispatcher;
 import com.google.common.collect.ImmutableMap;
 import java.util.Arrays;
 import java.util.Map;
