+++ core/src/test/java/com/datastrato/gravitino/listener/api/event/TestSchemaEvent.java
@@ -13,10 +13,10 @@ import static org.mockito.Mockito.when;
 import com.datastrato.gravitino.NameIdentifier;
 import com.datastrato.gravitino.Namespace;
 import com.datastrato.gravitino.catalog.SchemaDispatcher;
-import com.datastrato.gravitino.catalog.SchemaEventDispatcher;
 import com.datastrato.gravitino.exceptions.GravitinoRuntimeException;
 import com.datastrato.gravitino.listener.DummyEventListener;
 import com.datastrato.gravitino.listener.EventBus;
+import com.datastrato.gravitino.listener.SchemaEventDispatcher;
 import com.datastrato.gravitino.listener.api.info.SchemaInfo;
 import com.datastrato.gravitino.rel.Schema;
 import com.datastrato.gravitino.rel.SchemaChange;
