+++ core/src/test/java/com/datastrato/gravitino/listener/api/event/TestTableEvent.java
@@ -12,10 +12,10 @@ import static org.mockito.Mockito.when;
 import com.datastrato.gravitino.NameIdentifier;
 import com.datastrato.gravitino.Namespace;
 import com.datastrato.gravitino.catalog.TableDispatcher;
-import com.datastrato.gravitino.catalog.TableEventDispatcher;
 import com.datastrato.gravitino.exceptions.GravitinoRuntimeException;
 import com.datastrato.gravitino.listener.DummyEventListener;
 import com.datastrato.gravitino.listener.EventBus;
+import com.datastrato.gravitino.listener.TableEventDispatcher;
 import com.datastrato.gravitino.listener.api.info.TableInfo;
 import com.datastrato.gravitino.rel.Column;
 import com.datastrato.gravitino.rel.Table;
