+++ core/src/test/java/com/datastrato/gravitino/listener/api/event/TestTopicEvent.java
@@ -13,10 +13,10 @@ import static org.mockito.Mockito.when;
 import com.datastrato.gravitino.NameIdentifier;
 import com.datastrato.gravitino.Namespace;
 import com.datastrato.gravitino.catalog.TopicDispatcher;
-import com.datastrato.gravitino.catalog.TopicEventDispatcher;
 import com.datastrato.gravitino.exceptions.GravitinoRuntimeException;
 import com.datastrato.gravitino.listener.DummyEventListener;
 import com.datastrato.gravitino.listener.EventBus;
+import com.datastrato.gravitino.listener.TopicEventDispatcher;
 import com.datastrato.gravitino.listener.api.info.TopicInfo;
 import com.datastrato.gravitino.messaging.Topic;
 import com.datastrato.gravitino.messaging.TopicChange;
